MODEXPW-48 - move bulk edit listener to listeners package
