+++ src/main/java/org/folio/dew/batch/bulkedit/jobs/updatejob/BulkEditUpdateUserRecordsJobConfig.java
@@ -2,6 +2,7 @@ package org.folio.dew.batch.bulkedit.jobs.updatejob;
 
 import org.folio.dew.batch.JobCompletionNotificationListener;
 import org.folio.dew.batch.bulkedit.jobs.JobConfigReaderHelper;
+import org.folio.dew.batch.bulkedit.jobs.updatejob.listeners.BulkEditUpdateUserRecordsListener;
 import org.folio.dew.domain.dto.User;
 import org.folio.dew.domain.dto.UserFormat;
 import org.springframework.batch.core.ItemWriteListener;
