+++ src/main/java/org/folio/dew/batch/bulkedit/jobs/updatejob/listeners/BulkEditUpdateUserRecordsListener.java
@@ -1,4 +1,4 @@
-package org.folio.dew.batch.bulkedit.jobs.updatejob;
+package org.folio.dew.batch.bulkedit.jobs.updatejob.listeners;
 
 import lombok.RequiredArgsConstructor;
 import org.folio.dew.service.BulkEditRollBackService;
